refactor: 회고 - 몰입 결과 컴포넌트화
